Change return type of `modified_since()` to float

+++ prerender/cache/base.py
@@ -8,5 +8,5 @@ class CacheBackend:
     def set(self, key: str, payload: bytes, ttl: int = None, format: str = 'html') -> None:
         raise NotImplementedError
 
-    async def modified_since(self, key: str, format: str = 'html') -> int:
+    async def modified_since(self, key: str, format: str = 'html') -> float:
         raise NotImplementedError
